+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+### Breaking Changes
+
+- User feedback API reworked:
+  - Feedback no longer needs to be associated with a specific event - the only required parameter is the user message
+  - `SentryUserFeedback` class replaced with `SentryFeedback`
+  - `CaptureUserFeedback` function in `SentrySubsystem` replaced with `CaptureFeedback`
+  - `CreateSentryUserFeedback` function in `SentryLibrary` replaced with `CreateSentryFeedback`
+- On Windows and Linux, `ToString` function of `SentryId` class now returns the ID without dashes
+
+### Features
+
+- Added new API for capturing user feedback ([#1051](https://github.com/getsentry/sentry-unreal/pull/1051))
+
 ### Dependencies
 
 - Bump CLI from v2.51.1 to v2.52.0 ([#1049](https://github.com/getsentry/sentry-unreal/pull/1049))